fix(Friend): DateTime 계산 로직을 Data 레이어에서 수행하도록 수정
- preview에서 DateTime 계산으로 렌더링이 안되는 문제 개선
