+++ Near/app/src/main/java/com/alarmy/near/data/mapper/FriendMapper.kt
@@ -13,6 +13,9 @@ import com.alarmy.near.network.response.AnniversaryEntity
 import com.alarmy.near.network.response.ContactFrequencyEntity
 import com.alarmy.near.network.response.FriendEntity
 import com.alarmy.near.utils.logger.NearLog
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+import java.util.Locale
 
 fun FriendEntity.toModel(): Friend =
     Friend(
@@ -25,8 +28,25 @@ fun FriendEntity.toModel(): Friend =
         anniversaryList = anniversaryList.map { it.toModel() },
         memo = memo,
         phone = phone,
+        isContactToday = lastContactAt?.isToday() ?: false,
         lastContactAt = lastContactAt,
+        lastContactFormat = lastContactAt?.contactFormat()
     )
+
+fun String.contactFormat(): String {
+    val inputFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
+    val outputFormatter = DateTimeFormatter.ofPattern("M월 d일")
+
+    val date = LocalDate.parse(this, inputFormatter)
+    return date.format(outputFormatter)
+}
+
+private fun String.isToday(): Boolean {
+    val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd", Locale.KOREA)
+    val targetDate = LocalDate.parse(this, formatter)
+    val today = LocalDate.now()
+    return targetDate == today
+}
 
 fun ContactFrequencyEntity.toModel(): ContactFrequency =
     ContactFrequency(
